project_4/requirements-all.txt
matplotlib==3.7.5
opencv_python_headless==4.9.0.80
Pillow==11.0.0
Requests==2.32.3
scikit_learn==1.3.2
streamlit==1.39.0
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
